+++ src/boxed.rs
@@ -50,6 +50,8 @@ mod tests {
     fn percore_boxed_slice() {
         static STATE: Once<PerCore<BoxedSlice, FakeCoresImpl>> = Once::new();
 
+        FakeCoresImpl::set_core_index(0);
+
         STATE.call_once(|| {
             let boxed_slice: BoxedSlice = repeat_with(|| ExceptionLock::new(RefCell::new(42)))
                 .take(4)
@@ -59,6 +61,7 @@ mod tests {
         });
 
         {
+            // SAFETY: There are no exceptions in the simulated environment of the tests.
             let token = unsafe { ExceptionFree::new() };
             assert_eq!(*STATE.get().unwrap().get().borrow_mut(token), 42);
             *STATE.get().unwrap().get().borrow_mut(token) += 1;
@@ -71,7 +74,10 @@ mod tests {
         static STATE: LazyLock<PerCore<BoxedSlice, FakeCoresImpl>> =
             LazyLock::new(|| PerCore::new_with_default(4));
 
+        FakeCoresImpl::set_core_index(0);
+
         {
+            // SAFETY: There are no exceptions in the simulated environment of the tests.
             let token = unsafe { ExceptionFree::new() };
             assert_eq!(*STATE.get().borrow_mut(token), 0);
             *STATE.get().borrow_mut(token) += 1;